pyproject.toml
[tool.poetry]
name = "kiwi-scp"
version = "0.2.0"
description = "kiwi is the simple tool for managing container servers."
authors = ["ldericher <40151420+ldericher@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.6.1"
attrs = "^21.2.0"
click = "^8.0.3"
pydantic = "^1.8.2"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
toml = "^0.10.2"
virtualenv = "^20.8.1"

[tool.poetry.scripts]
kiwi = "kiwi_scp.scripts.kiwi:main"
kiwi_next = "kiwi_scp.scripts.kiwi_next:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
